refactor: split shared/types/novel.ts into domain subdirectory (#8)
Split the 939-line novel.ts into shared/types/novel/ with four
domain files (core, chapter, volume, workflow) plus a barrel
index.ts that re-exports everything. External import paths remain
unchanged since the package exports wildcard resolves directories.

- core.ts (309 lines): Novel interface, base enums, pipeline, model route, story state
- chapter.ts (250 lines): Chapter, ChapterEditor* types
- volume.ts (304 lines): VolumePlan*, strategy, beat sheets, storyline, StoryPlan
- workflow.ts (89 lines): Replan, audit, creative decisions, plot beats
- index.ts (78 lines): barrel re-exports including external module pass-throughs
